Fixing ft_strspn return type

+++ src/str/ft_strspn.c
@@ -6,7 +6,7 @@
 /*   By: cacharle <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/02/10 04:29:13 by cacharle          #+#    #+#             */
-/*   Updated: 2020/02/10 04:33:11 by cacharle         ###   ########.fr       */
+/*   Updated: 2020/08/27 09:56:04 by charles          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strspn(const char *s, const char *charset)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (ft_strchr(charset, s[i]) != NULL)
